[cluster-statistics] Adds operation retrieve api

+++ rest/1.0/common/v1/pdu/cluster_request.go
@@ -1,5 +1,5 @@
 package pdu
 
 type ClusterRequest struct {
-	TimeoutSec int32 `json:"timeout_sec,omitempty"`
+	TimeoutSec uint16 `json:"timeout_sec,omitempty"`
 }
